Freeze the query of the results
Before this, we plot the current selected series along with the latest
result. With this, the query on the result chart is changed as we
select a new query, which does not make sense.
Now, we can freeze the query of the result and it won't change until
we perform a new search.